Consolidate `chat-with-code` page into `chat` page, removing the former and enhancing the chat interface with features like file attachments, model selection, and web search.
