+++ .husky/pre-commit
@@ -1,2 +1,2 @@
-npm run lint
 npm run format
+npm run lint
